Se actualiza el proyecto a Jakarta
Se sube la versión de Java a 17
Se corrigen los imports y ns para utilizar Jakarta
Se cambia la paqueteria a com.mihtan.mailtester
Se actualiza el Readme
Se remplaza log4j por logback
Se remplaza el plugin de Payara por Cargo
