+++ README.md
@@ -1,2 +1,3 @@
 # java-mail-tester
-Utilitario para probar una configuración de Email con el API de Java y Apache Commons Email.
+
+Utilitario para probar una configuración de Email con el API de Jakarta.
