+++ pom.xml
@@ -2,129 +2,82 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>com.ejemplo</groupId>
-    <artifactId>Mail-Tester</artifactId>
-    <version>1.0.0</version>
+    <groupId>com.mihtan</groupId>
+    <artifactId>mailtester</artifactId>
+    <version>2.0.0</version>
     <packaging>war</packaging>
 
-    <name>Mail-Tester</name>
+    <name>MailTester</name>
 
     <properties>
-        <endorsed.dir>${project.build.directory}/endorsed</endorsed.dir>
+        <jakartaee-api.version>10.0.0</jakartaee-api.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+        <project.report.sourceEncoding>UTF-8</project.report.sourceEncoding>
+        <maven.compiler.release>17</maven.compiler.release>
+        <maven.compiler.source>17</maven.compiler.source>
+        <maven.compiler.target>17</maven.compiler.target>
+        <logback.version>1.3.15</logback.version>
+        <cargo.version>1.10.15</cargo.version>
+        <payara.version>6.2025.1</payara.version>
     </properties>
 
     <dependencies>
         
-        <!-- :::::::::::::::::::::::::::::::::::::::::::::: -->
-        <!-- ::::::::::::   JAVA EE API   ::::::::::::::::: -->
-        <!-- :::::::::::::::::::::::::::::::::::::::::::::: -->
+        <!-- Jakarta EE Platform -->
         <dependency>
-            <groupId>javax</groupId>
-            <artifactId>javaee-web-api</artifactId>
-            <version>6.0</version>
+            <groupId>jakarta.platform</groupId>
+            <artifactId>jakarta.jakartaee-api</artifactId>
+            <version>${jakartaee-api.version}</version>
             <scope>provided</scope>
         </dependency>
-        
-        <!-- :::::::::::::::::::::::::::::::::::::::::::::: -->
-        <!-- ::::::::::::  COMMONS EMAIL  ::::::::::::::::: -->
-        <!-- :::::::::::::::::::::::::::::::::::::::::::::: -->
+
+        <!-- SLF4J -->
         <dependency>
-            <groupId>org.apache.commons</groupId>
-            <artifactId>commons-email</artifactId>
-            <version>1.5</version>
+            <groupId>ch.qos.logback</groupId>
+            <artifactId>logback-classic</artifactId>
+            <version>${logback.version}</version>
         </dependency>
 
-        <!-- :::::::::::::::::::::::::::::::::::::::::::::: -->
-        <!-- ::::::::::::::    LOGGING    ::::::::::::::::: -->
-        <!-- :::::::::::::::::::::::::::::::::::::::::::::: -->
-        <dependency>
-            <groupId>org.slf4j</groupId>
-            <artifactId>slf4j-api</artifactId>
-            <version>1.7.9</version>
-        </dependency>
-        <dependency>
-            <groupId>org.slf4j</groupId>
-            <artifactId>slf4j-log4j12</artifactId>
-            <version>1.7.9</version>
-        </dependency>
-
-        <!-- :::::::::::::::::::::::::::::::::::::::::::::: -->
-        <!-- ::::::::::::::  BootsFaces   ::::::::::::::::: -->
-        <!-- :::::::::::::::::::::::::::::::::::::::::::::: -->
+        <!-- BootsFaces -->
         <dependency>
             <groupId>net.bootsfaces</groupId>
             <artifactId>bootsfaces</artifactId>
-            <version>1.4.2</version>
+            <version>2.0.1</version>
             <scope>compile</scope>
+            <classifier>java17</classifier>
         </dependency>
         
     </dependencies>
 
     <build>
+        <finalName>mail-tester</finalName>
         <plugins>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-compiler-plugin</artifactId>
-                <version>2.3.2</version>
-                <configuration>
-                    <source>1.6</source>
-                    <target>1.6</target>
-                    <compilerArguments>
-                        <endorseddirs>${endorsed.dir}</endorseddirs>
-                    </compilerArguments>
-                </configuration>
+                <version>3.13.0</version>
             </plugin>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-war-plugin</artifactId>
-                <version>2.1.1</version>
+                <version>3.4.0</version>
                 <configuration>
                     <failOnMissingWebXml>false</failOnMissingWebXml>
                 </configuration>
             </plugin>
             <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-dependency-plugin</artifactId>
-                <version>2.1</version>
-                <executions>
-                    <execution>
-                        <phase>validate</phase>
-                        <goals>
-                            <goal>copy</goal>
-                        </goals>
-                        <configuration>
-                            <outputDirectory>${endorsed.dir}</outputDirectory>
-                            <silent>true</silent>
-                            <artifactItems>
-                                <artifactItem>
-                                    <groupId>javax</groupId>
-                                    <artifactId>javaee-endorsed-api</artifactId>
-                                    <version>6.0</version>
-                                    <type>jar</type>
-                                </artifactItem>
-                            </artifactItems>
-                        </configuration>
-                    </execution>
-                </executions>
-            </plugin>
-            <plugin>
-                <groupId>fish.payara.maven.plugins</groupId>
-                <artifactId>payara-micro-maven-plugin</artifactId>
-                <version>1.0.1</version>
+                <groupId>org.codehaus.cargo</groupId>
+                <artifactId>cargo-maven3-plugin</artifactId>
+                <version>${cargo.version}</version>
                 <configuration>
-                    <payaraVersion>5.184</payaraVersion>
-                    <deployWar>false</deployWar>
-                    <commandLineOptions>
-                        <option>
-                            <key>--autoBindHttp</key>
-                        </option>
-                        <option>
-                            <key>--deploy</key>
-                            <value>${project.build.directory}/${project.build.finalName}</value>
-                        </option>
-                    </commandLineOptions>
-                    <contextRoot>/</contextRoot>
+                    <container>
+                        <containerId>payara</containerId>
+                        <artifactInstaller>
+                            <groupId>fish.payara.distributions</groupId>
+                            <artifactId>payara</artifactId>
+                            <version>${payara.version}</version>
+                        </artifactInstaller>
+                    </container>
                 </configuration>
             </plugin>
         </plugins>
