+++ src/main/webapp/WEB-INF/payara-web.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE payara-web-app PUBLIC "-//Payara.fish//DTD Payara Server 4 Servlet 3.0//EN" "https://docs.payara.fish/schemas/payara-web-app_4.dtd">
+<payara-web-app error-url="">
+    <context-root>/mail-tester</context-root>
+    <!-- false=usar clases del server, true usar clases del war -->
+    <class-loader delegate="true"/>
+</payara-web-app>
