+++ src/main/webapp/index.xhtml
@@ -1,10 +1,10 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE html>
 <html xmlns="http://www.w3.org/1999/xhtml"
-      xmlns:f="http://java.sun.com/jsf/core"
-      xmlns:h="http://java.sun.com/jsf/html"
+      xmlns:f="jakarta.faces.core"
+      xmlns:h="jakarta.faces.html"
       xmlns:b="http://bootsfaces.net/ui"
-      xmlns:ui="http://java.sun.com/jsf/facelets">
+      xmlns:ui="jakarta.faces.facelets">
 <h:head>
     <title>Mail Tester</title>
 </h:head>
